Refactoring complete towards the ServiceBusAdapter
